pyproject.toml
[project]
name = "code-review-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.136.0",
    "langchain-core>=1.3.0",
    "langchain-groq>=1.1.2",
    "langgraph>=1.1.9",
    "pygithub>=2.9.1",
    "python-dotenv>=1.2.2",
    "tavily-python>=0.7.23",
    "uvicorn>=0.45.0",
]
